change int to long, amount to balance

+++ src/main/java/com/bank/jdt/RESTSaving/Entity/Saving.java
@@ -21,7 +21,7 @@ public class Saving implements Serializable {
     private long customerId;
     @Column(name = "account_saving")
     private long accountSaving;
-    private int balance;
+    private long balance;
     @Column(name = "is_active")
     private boolean isActive;
     @ApiModelProperty(hidden = true)
